more sophisticated Errors

+++ src/container.js
@@ -1,3 +1,4 @@
+const { IndexValidationError } = require('./validation-error');
 const { Record } = require('./core/record');
 const { Compartment } = require('./core/compartment');
 const { Species } = require('./core/species');
@@ -35,7 +36,10 @@ class Container {
   }
   insert(q){
     // check
-    expect(q).to.have.property('id').with.a('string');
+    if(!q)
+      throw new IndexValidationError(q);
+    if(!q.id || (typeof q.id !== 'string'))
+      throw new IndexValidationError({id: q.id});
     expect(q).to.have.property('class').with.a('string');
     // check if class is in the list
     let selectedClass = this.classes[q.class];
@@ -50,8 +54,11 @@ class Container {
     return simple;
   }
   update(q){
+    if(!q)
+      throw new IndexValidationError(q);
+    if(!q.id || (typeof q.id !== 'string'))
+      throw new IndexValidationError({id: q.id});
     expect(q).not.to.have.property('class');
-    expect(q).to.have.property('id').with.a('string');
     let index = q.space ? (q.space + '.' + q.id) : q.id;
     let targetComponent = this.storage.get(index);
 
@@ -73,8 +80,11 @@ class Container {
     }
   }
   delete(q){
+    if(!q)
+      throw new IndexValidationError(q);
+    if(!q.id || (typeof q.id !== 'string'))
+      throw new IndexValidationError({id: q.id});
     expect(q).not.to.have.property('class');
-    expect(q).to.have.property('id').with.a('string');
     let index = q.space ? (q.space + '.' + q.id) : q.id;
     let targetComponent = this.storage.delete(index);
